Rebuilt class to run as method
Made some minor changes to make it easier to run the code as a loop. 

Running the parser from a main loop provides much better results when using alternative keywords. If you are parsing for a large set of alternative keywords, I recommend you structure your search as I detail in the below issue. See the master.py file in my memeClassifier project for the entire implementation 

https://github.com/hardikvasa/google-images-download/issues/25